Fix #2336 for TextElementEnumerator
There was a lot of duplication in this test suite again. Some of the
test code is duplicated, and also similar but less in depth than
StringInfo.GetTextElementEnumerator.

Eliminating duplicate code allowed me to put all of this into one file -
even though I was told off before for combining test files into a single
file, this approach is necessary for a clean test suite, and was not done for the sake of combining tests into a single file.