Fix missing types from S.Globalization.Calendars
System.Globalization.Calendars should be a facade assembly, but in
030e29992f6be20ffb9d9d3e799118f3354bce22 we accidently made it facade
only for net46.

We didn't catch this during testing because the test assembly was always
running against the package version, but in Helix, where we test against
packages all the time, we hit the issue.

Make the assembly a partial facade all the time and update the test
project so we actually test agains the live code (doing so caused the
test to fail before the above fix, so it lets us know we won't miss this
issue in the future).