Move RuntimeInformation name detection into System.Native
We currently build a managed .dll for each supported platform, purely for the purpose of a single string denoting the target platform's name.  Since we already need to build System.Native per platform anyway, we can just expose the name from there.  With this change, a single System.Runtime.InteropServices.RuntimeInformation.dll can be shared by all flavors of Unix, and in the future we won't need to specialize it or its packages to bring up additonal platforms.