Fix Console.ReadKey() backspace handling on OSX
The terminfo files for xterm* on OSX contain ```^H``` for the kbs entry (Backspace key); this can be seen, for example, with ```infocmp -1 xterm-256color | grep kbs```.  However, the terminal is sending ```^?```.  This commit addresses the discrepancy by consulting the termios control codes and using them as an override for what's in termios.