Increase linger time in SendRecvAsync tests
These tests are failing with an OperationAborted status code on OSX, very rarely.  I can't reproduce the problem locally, even if I try to artifically induce the failure.  I'm guessing that the timing works out differently on the CI machines, making this (somewhat) more likely, and the only legitimate case I can see where this would happen is if the client closes the TCP connection before the server has finished reading all bytes.  Increasing the linger time should help with that, so let's give it a try.