Remove the issue and PR templates
We can replace them later with guidance, examples, etc.   Any such thing would need to be in a different form (e.g. the rendered view of these .md files is very different from the content, where a lot of the guidance is in comments that don't render), so I'm simply deleting these for now.