Add a test for ctrl-C handling on Unix
I'd previously added one, but then removed it because it involved sending a ctrl-C to the current process, which would cancel the remainder of the xunit tests in the process.  This commit reinstitutes the test, but in a manner that does the interesting work in a separate process so that it's canceled rather than xunit getting canceled.