Add Debug.Assert messages to System.Net.Sockets Unix code
With the current lack of PDB support on Unix, Debug.Assert failures don't include file / line number information, which makes it harder to debug, especially when there are multiple asserts in a single function.  This commit adds messages to all of the Debug.Asserts in the Unix-specific code in System.Net.Sockets.