Fix SocketException error codes and messages on Unix
SocketException's public ctor takes an int, which it treats as as both a SocketError and the native error code, passing it down to the base Win32Exception to get the platform-specific message for the error.  However, SocketError is only the same as the native error code on Windows; on Unix, a translation is required, but that translation is missing.  As a result, every SocketException thrown from System.Net.Sockets on Unix comes out with an unhelpful message "Unknown error 12345", where 12345 is the value for the SocketError, and the base native error code is set inappropriately as the SocketError's value rather than as the real errno value.

This change puts a reverse mapping in place, such that the SocketError provided to SocketException is then translated back (as best as possible) into the underlying platform's native error code, which is then both stored on the base exception and used to get the appropriate platform-specific error message. So, for example, rather than getting an error like:
```
System.Net.Sockets.SocketException: Unknown error 10047
```
you now get an error like:
```
System.Net.Sockets.SocketException: Address family not supported by protocol
```

Additionally, several SocketErrors map to errno values that weren't included in our Interop.Errors enum.  I've added them.

And, SocketException had dead code in the form of an "internal SocketException(EndPoint endPoint)" ctor that was never used.  As a result, the private _endPoint field was always null, and the Message property would always just return base.Message.  I've removed the constructor, the field, and the Message property override.  SocketException also had an ErrorCode property in the implementation that wasn't in the ref contract; I've removed it from the implementation, too.