System.Net.Sockets: Socket: Support setting MulticastInterface socket option to interface index on non-Windows platforms
The MulticastInterface option can be set to the IPv4 address of an outgoing interface.
On Windows, the network interface for outgoing multicast traffic can also be set to an interface index like this:
	socket.SetSocketOption(SocketOptionLevel.IP, SocketOptionName.MulticastInterface, IPAddress.HostToNetworkOrder(interfaceIndex))

This works because Windows detects when the address passed as IP_MULTICAST_IF is in the 0.x.x.x block. When that is the case, the address is interpreted as an index.

On non-Windows platforms, an in_addr passed to IP_MULTICAST_IF is always interpreted as an address. Since there will never be an interface with a 0.x.x.x address, the call always fails. Linux and OSX have support for setting the multicast interface by making use of an ip_mreqn instead of an in_addr.

This patch detects when the address is in the 0.x.x.x block and makes an ip_mreqn request.