Removing an assert in BitVector32
This assert came over from the full framework source.  This should really have been a check-and-throw-exception, as the assert will be compiled out in release code and this is a public method.  We can't add that exception now, and we have tests that are explicitly calling this method, which is triggering debug assert failures and causing unnecessary failure details in our CI logs.  I'm just removing it.