Remove special EOF handling in SocketAsyncContext event loop
This code makes the assumption that if the other end of a socket is closed, all reads that will ever be issued on this end are already pending, so it's ok to shut down the read queue.  This is not necessarily true; we need to allow other threads to continue to issue reads until they've actually read all data.

This is resulting in assert failures (#6506) and is a likely cause of erroneous OperationAborted failures (#5750).

It's not clear to me what value the EOF handling is providing currently; currently, aside from erroneously shutting down the queue, the only difference from the normal read event logic is that we also unregister from read events.  Since we no longer are able to free handles, etc, on unregistration, it's not clear that we get any benefit from unregistering an event we'll probably never receive again.

This change therefore simply removes this logic.