CredentialCache.Remove compat: Don't throw when key not found
This commit changes `CredentialCache.Remove` to not throw in .NET
Core, matching the behavior of the full framework.

In the full framework, `CredentialCache.Remove` does not throw when
the key isn't found, which is consistent with the behavior described
in the MSDN documentation:

> If authType is null or uriPrefix is null, or no matching credential
is found in the cache, this method does nothing.