Add a redist package for InteropServices
Currently anything depending on Interop get's baselined to the latest
which takes a dependency on System.Runtime 4.1 which breaks the
baseline.  Additionally this was creating downgrades when folks
referenced both S.R 4.0.20 or lower and S.R.IS.

Once I have this package I'll need to update the baseline in buildtools.