Remove wrongly hard-coded /usr/pkg/include path for NetBSD
When I started to work with CoreFX it already had wrongly hard-coded optional
path for pkgsrc dependencies it may be /usr/pkg/include, but it's purely
optional path. Also pkgsrc is purely optional on NetBSD (and any other system,
even if it's the default package manager like in SmartOS).

The correct way of handling it via CMake is finding includes (there was a macro
for it) or discovering packages. I went for the latter.