Minor tweaks to Dataflow's ImmutableList
In one place, Dataflow has need for a minimal immutable array data structure.  I don't think it makes sense to take a dependency on System.Collections.Immutable.dll just for that, but there are a few minor cleanups that can be done to the implementation in dataflow to bring it a bit closer in design to the "real" one, e.g. making it a struct, renaming it from ImmutableList, etc.