Add a few PlatformNotSupported exception messages
Most of the PlatformNotSupportedExceptions we throw in our Unix code paths are unconditional, e.g. in System.Console:
```C#
public static string Title
{
    get { throw new PlatformNotSupportedException(); }
    set { ... }
}
```
making it easy to understand without a customized message that the method being called isn't supported.

However, when there's more nuance, e.g. when the operation isn't supported because of a particular parameter value being used, we do want to use a message to clarify exactly what's not supported.  Most such cases already have a custom message, but a few don't.

This commit fixes those that I could find that still need one (if we encounter more, we can add those on a case-by-case basis).