Remove S.Runtime 4.1.0 workarounds
Previously we were incorrectly lifting InteropServices references to
4.1.0.0 in the packages  https://github.com/dotnet/corefx/pull/6648.

This caused an issue when consuming those packages since it would
cause a typeforward to System.Runtime and we weren't referencing
the version of System.Runtiem to satisfy the forward.  Now that I've
fixed the incorrect baselining we can remove the System.Runtime
workarounds.