Return correct SslStream.KeyExchangeValue when algorithm is ECDH
We needed to add another constant for the ECDH key exchange algorithm, to match the behavior on Windows.  Also, the string comparisons used when parsing out the results for `CryptoNative_GetSslConnectionInfo` were not doing the right thing; they needed to compare the *lengths* of the strings as well, to avoid, for example, "ECDH" matching "ECDH/RSA".