Fix System.Data.SqlClient package
A couple problems with this package that surfaced after the change to
the version.  The net45 facade had an incorrect version.  There was no
net46 facade.  I've fixed up both of these problems and moved the 4.0
reference assembly to a folder that matches our conventions for
semantic-versioned reference assemblies (3-part is only used for the
legacy versions that didn't follow semantic versioning).