Fix default permissions in UnixFileStream
We currently have UnixFileStream create new files with the permissions 700.  This doesn't map to what's commonly expected on Unix, where most utilities default instead to 666.

This commit changes the default to be 766, which attempts to map to what's commonly expected on Unix, while also factoring in the default on Windows, where a new file can be executed by default.  All of these permissions are of course subject to the system's umask, such that the actual permissions applied will likely be less than what's specified.

The commit also fixes up some code in X509Certificates, which was coded to rely on the more limited permissions being applied.