Add ENOENT to SocketErrorPal
Unix domain sockets can result in ENOENT failures due to the specified file path not existing.  There's no corresponding entry in SocketError, so it often ends up showing up as an unknown -1 error.  This commit just adds a mapping to the closest entry possible, AddressNotAvailable, making it both a bit easier to diagnose and making it possible to filter based on AddressNotAvailable in consuming code.