Fix incorrect asserts in Socket.ConnectAsync for Unix domain sockets
A previous change allowed AddressFamily.Unix and non-IPEndPoints to be used in System.Net.Sockets, removing some asserts and changing some if/elses to support it.  One case was missed, though, in ConnectAsync, which causes asserts and runtime failures when using Socket.ConnectAsync with Unix domain sockets.  This commit just makes the same change to that code path that was made to other similar ones.