Implement NamedPipe*Stream on Unix on Unix domain sockets
Today, NamedPipeServer/ClientStream are implemented on top of FIFOs, aka named pipes.  There are some unfortunate limitations to this that are causing problems for typical usage of NamedPipeServer/ClientStream:
- FIFOs can be constructed as both read and write (i.e. InOut), but a writer can then immediately read what was written.  In other words, whereas one might expect there two be two data channels, one in each direction, there's only one that everyone can read and write from.
- FIFOs can be connected to by any number of readers and writers, which has the effect of meaning a reader doesn't unblock when a writer disconnects, because there may be another writer in the future.

To address these significant limitations, this commit moves NamedPipeServer/ClientStream to be built instead on top of Unix domain sockets.  Unix domain sockets are an IPC mechanism for local communication using the sockets model, and we have support for it via System.Net.Sockets.

Both of the aforementioned problems are addressed by this change.  A few other positive outcomes:
- The participants on each side of the connection can get the credentials of the other side, so GetImpersonatedUserName is now implemented.
- Our Socket implementation has a good async implementation, so we can use it for all of the async operations exposed rather than queueing a work item that then blocks doing the synchronous operation.

There are a few downsides, but they're worth the tradeoff:
- FIFOs have the same general blocking behavior as anonymous pipes and as named pipes on Windows: a call to Write doesn't complete until an associated Read comes along.  That's not the case with Unix domain sockets, where a Write typically just deposits the data into a buffer.  Writes on a socket only block when the buffer is full.
- Unix domain sockets are considered to be a bit slower than FIFOs.
- Changing the buffer size of a FIFO impacts the whole FIFO, whereas changing the buffer size for a socket impacts just that end of the socket.

There continue to be several features that we don't support, e.g. message transmission, max number of servers, etc.