Don't get available bytes for the socket unless we actually need the result
While reading through the code I noticed that we're making unnecessary `ioctl` calls in Receive calls, to get the count of available bytes.  This turns out to only be used if the caller supplied a list of buffers, which is not the typical case.  This change moves the `ioctl` call so that it happens only if needed, which should help with throughput in most scenarios.