Implement SocketPal.Poll on Unix with poll instead of select
SocketPal.Poll on Unix is currently implemented with select.  This is causing problems due to limitations of select, which fails with file descriptors >= FD_SETSIZE.  I've changed the implementation to use poll instead, which both fixes the problem and results in simpler code.  I've also added a few tests, and in the process of doing so cleaned up some related tests to use theories.