Implement exception filters in interpreterd expressions.
Fixes #3841.

Finding which catch block to run in a try-catch now happens entirely
in a filter in the EnterTryCatchFinallyInstruction, where previously
all exceptions would be caught and the matching occur within that
catch-block.

As well as allowing for filter support, this also means that there is
no catch if there was no corresponding catch block (all catch blocks
caught different exception types), which makes the exception-handling
of the interpreter better match the stack-walk of the CLR.