Fix MMF test failure due to inheritable handles
Some tests create maps with inheritable handles.  Other tests launch processes.  If those tests end up running concurrently in a manner where the launched process inherits the handle, it'll end up extending the lifetime of the map beyond what the tests were expecting.  This has been leading to random test failures like "Cannot create a file when that file already exists".

This commit simply disables parallelization of the MMF tests.  This has little impact on the running time of the tests, but if in the future we add the ability to disable inheriting handles when using Process.Start, we could use that instead.