Fix unix System.Private.Uri package
Fixes https://github.com/dotnet/corefx/issues/6883

When we turned off OSGroup filtering we didn't pass the correct
OSGroup from the pkgproj so we ended up with the default OSGroup
which is windows and so we ended up with a window version of this
library in runtime.unix.System.Private.Uri package. Which means
unix didn't have a version of this library.