Remove Title_set test method
This method is failing intermittently in the outerloop automated testing
which can't be reproed locally. Interestingly we have only seen this call
failing in Title_set test method and not in Title test method both
of which call set_Title. For now, removing this method since we get the same
coverage in both the test methods.