Added support to map particular DiagnosticSource Events to particular EventSource Events.
This is to support activities in Diagnostic Source.

By default, all DiagnosticSource Events are forwarded to the EventSource by a EventSource event
called 'Event'.    However in order for DiangosticSources to participate in EventSource 'start-stop'
activities, you need the EventSourceEvents to be marked as being starts and stops.

This change adds a 6 EventSource Events that are so marked (e.g. Activity1Start, Activity2Stop ...)
and adds a new  @EventSourceEventName syntax when specifying a FilterAndTransformSpec.

    DiagnosticSourceName/EventName@EventSourceEventName

This allows you to specify that paritcular DiagnosticSource Events map to particular EventSource
event and thus participate in start-stop activity tracking.