Distinguish raw IPv4 from IPv6 in unix Ping implementation
Some systems enable raw sockets to be opened for ICMP but not ICMPv6, but our current test in Ping for whether to open a raw socket directly is based only on ICMP, which then causes run-time failures when an IPv6 address is used.  This commit augments the current check to distinguish between the two.