Change http shim to support getting better version info
Currently the shim exposes a single method to get just the information that was needed when this was written.  Every time we need to get additional version info, we need to change the existing signature (which won't version well) or add additional similar signatures (will which be confusing).  Instead, I've changed it to have more fine-grained entrypoints, each exposing a particular piece of info.  We may end up making a few more interop calls on first-use, but the overhead is negligable.

With that scheme, I've then exposed additional entrypoints for getting version information, in order to provide better tracing and exception error messages.

This is then also used in tests to only run tests when certain features are available.