Baseline Win7 CNG symmetric encryption tests
AesCng and TripleDESCng both do their cryptographic operations via NCrypt
exclusively.  On Windows 7 the software key provider does not support
symmetric encryption, so this fails.

As this failure is too large to address for RC2, and the test material is in a
complex location to disable for just one OS version, make the CNG symmetric
tests use an alternative implementation for now, to get the badge green and
help track regressions.