Fixes #7126
When building on OSX, the msbuild AssignCulture task miscategorizes some resources. See #541
This change removes dots from the name of the embedded resource, which is a workaround for the msbuild issue.

The change also adds a check for null stream, to align with what other components do and to help in diagnostics.

It also changes the type used to retrieve the resource to a public type so that it works when the runtime does not
support private reflection.