Fix OSX-specific issue from merge of "legacy" tests
Some tests brought over from "legacy" used the SocketUdpServer type, which already existed in the main Sockets test assembly, but was implemented slightly differently.  It used ReceiveFrom, which does not work on dual-mode sockets on OSX.

Updated this to use Receive instead.  This does not affect test coverage, as this is used purely as a helper for tests that are exercising various Send* methods; the Receive* methods are covered in other tests.