Fix and re-enable XmlConvert test
The Windows issue was already closed, but the test was still deactivated for it.  The Unix issue is due to timezone information differing for the ancient date being used; this is "by design", and since this is a test of formatting rather than calendars and timezones, I've simply changed to a more recent date where the timezone information matches.