Reduce allocations in CookieContainer.GetCookieHeader
HttpClient defaults to UseCookies=true, which means by default every request ends up asking the CookieContainer to get the cookie header that should be used. Today, even if there aren't any cookies for the Uri, this ends up allocating a CookieCollection, a StringBuilder, a couple of List<T>s, an iterator, etc., none of which are actually necessary.  And if there are cookies for the Uri, the serialization of each cookie results in a handful of strings, char[]s, and string[] allocations. This commit fixes the implementation to avoid these. (There are still some that could be cleaned up with further work, but it becomes much more invasive.)