Refactor HttpClient with async/await
This commit addresses just the HttpClient class and not the rest of the library, but those can follow.  When HttpClient was initially written, async/await didn't exist, and it was written to use ContinueWith.  By switching to async/await, we can both improve the readability and maintainability of the code and improve performance.