Improve serial throughput of HttpClient on Unix
HttpClient on Unix has a thread dedicated to multiplexing the processing of any number of concurrent requests on that client.  When a request comes in and that worker thread isn't running, a worker is spun up, and that worker will service any number of requests that come in while it's up.  Then, when requests quiesce and there's no more processing to be done, the worker goes away.  If more work later arrives, a new worker is spun up.

This is done to avoid burning a thread per HttpClient when there's no work to be done on that client.  It works well for bursts of activity as well as for sustained concurrent usage.  However, if the client is used for a sequence of requests, one after the other, there's a good chance that the worker will go away between requests, requiring a new worker to be spun up for each.  This has non-trivial costs, both in terms of CPU processing and in terms of managed and native resource allocation/deallocation.

This commit enables a worker to briefly pause before exiting, blocking for a short period of time for new work to arrive rather than simply exiting.