Add an Online and Offline revocation mode test
This test may result in contacting the network to download a CRL, or
intermediate certificates, so it is being confined to OuterLoop.

This doesn't test that revocation actually works.  Unfortunately that is hard to
build into a unit test due to the non-constant nature of time: once a certificate
has gone out of validity a CA has no obligation (other than their own policy)
to keep a revoked certificate on the CRL.

The best way to write such a test is to use a live source of truth for a revoked
certificate; but the notion of "live" is subject to change.

An alternate mechanism would be to create a private CA with an
unfathomably large expiration time, issue and subsequently revoke a
certificate, and use that certificate, CA, and published CDP as a controlled
portion of the test.  Such an infrastructure dependency is beyond the scope
of this change (and would still fail for very old snapshots given enough time).