Fix generation of IS.PInvoke contract
Right now the new version of IS that typeforwards to this is NS1.5
this must also be NS1.5 to avoid type collisions.

As a seperate follow up issue we need to come up with a strategy to
oob this contract all the way back to NS1.1 so that we don't have COM
as part of NETStandard.Library.