Ignore EINVAL from SO_LINGER on OSX
On OSX, SO_LINGER can return EINVAL if the other end of the socket is already closed.  In that case, there is nothing for this end of the socket to do, so there's no reason to "linger."  Windows and Linux do not return errors in this case, so we'll simulate success on OSX as well.