Remove <UnsupportedPlaforms> from run-test.sh
Since #6572 was merged, we only build test projects for the platforms
they test.  This means the UnsupportedPlatforms metadata is no longer
needed (instead using the .builds file to control when the project is
built is the right thing to do).

Leverage this in run-test.sh, we now simply loop over all the tests in
the relevent folders under bin/tests, depending on the platform in
question.