Remove CultureInfoConverter
This type converts English culture display names to CultureInfo objects and vice versa (e.g. new CultureInfo("ru") <--> "Russian"). To accomplish this it relies on being able to enumerate the execution environment's supported cultures. That's impossible in corefx because the framework relies in turn on the operating system for such information, and some platforms, e.g. Linux and Windows 10, do not provide it. See #1669 for more context.