Extend DNS cache timeout in CurlHandler
Now that we're allowing the multi handle to live for as long as the CurlHandler, the DNS cache timeout starts to matter much more.  libcurl will cache the results of DNS lookups, and by default will timeout entries after 1 minute.  Our Windows implementation defaults to 10 minutes, so I'm extending our Unix implementation to match that.