Improve reliability of CurlHandler MultiAgent worker
If something unexpectedly goes wrong in our worker loop, and in particular when queueing a replacement worker, we may end up leaving a request sitting in the request queue with no one to process it.  This commit ensures that if something does go horribly wrong, the agent is left in a state where subsequent incoming requests can still kick off another worker, and in the meantime any pending requests in the queue are failed, so that code blocking on them doesn't simply hang.

As I was looking at this and some of our dispose usage, I also cleaned up a few places that could be simplified using the C# 6 ?. syntax.