Add a TestWithLocalLibraries code coverage build
Our code coverage run doesn't currently employ /p:TestWithLocalLibraries=true.  While that's good for the common need of understanding how tests for a particular library are exercising that library, it would also be good to be able to understand how the libraries are exercised by any tests and other libraries in the repo.  This commit adds an additional such job.