Fix Unix X509Store directory permissions check
Users who create their first X509Store on build 23910 or later will get an exception that the directory has the wrong permissions. No test existed at the
time, and manual testing failed to identify the distinction between "continued
use" and "initial creation".

This changes the permissions check to ensure that the user permissions are
correct, and ignores the group and other permissions.  The files are explicitly
tested for rw?--?--?, and set to rw------- on mismatch, so there is no major
concern with the directory being overly readable.  It also adjusts the exception message to match the adjusted requirement, and adds a test.

This test will still suffer from the x509stores directory already existing on
established machines, since otherwise that runs the risk of breaking the
user's my and root stores if test cleanup goes awry.