Add checked coreclr cross platform outerloop runs
It would be nice to get coverage where we use debug versions of both
CoreFX and CoreCLR so we can get coverage with asserts
enabled (especially given that some parts of CoreFX are implemented in
native and managed code in CoreCLR).

We use checked CoreCLR, which has some compile time optimizations
applied (so runs should hopefully complete in a reasonable amount of
time) but still retains all the asserts.