Fix Process.ExitCode when used in conjunction with EnableRaisingEvents
Our Process waiting/events implementation on Unix is complex due to the contraints of the underlying platform, e.g. you can only waitpid on a child once and then its state can be reclaimed by the OS, you can only waitpid on a child process, etc.  As a result, we maintain a shared table mapping process IDs to wait state information about that process.  To avoid leaks, we employ a reference counting system combined with dispose/finalization, such that entries are removed from the table when all consumers of the information have unregistered.  Exit code information is part of that wait state.

Today when you EnableRaisingEvents, we create and populate that table with wait state information, which will be filled in when the process exits.  However, if no one else has add-ref'd that wait state information, it'll then be removed from the table.  If someone else then comes along and asks for the wait state information, a new one is manufactured, and at this point we're unable to get some of the wait-related info, like exit code, because it can only be obtained from the underlying system once.  This means if you EnableRaisingEvents, allow the process to exit, and then after some period of time ask for the process' exit code, you'll likely get back 0 rather than the actual code.

The fix is to ensure that the Process itself maintains a ref on its own wait state information.  We already do this, but we only initialize that ref when waiting on the process, when asking whether its exited, etc.  We instead need to ensure it's done as early as possible, and the earliest it's possible is as soon as we have an ID for the process.